backend/theory-service/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
click==8.1.8
ecdsa==0.19.1
fastapi==0.115.12
h11==0.14.0
idna==3.10
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
pyasn1==0.4.8
pydantic==2.11.0
pydantic-settings==2.8.1
pydantic_core==2.33.0
pytest==8.3.5
python-dotenv==1.1.0
python-jose==3.4.0
redis==5.2.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.34.0
